Module Build_info.V1

module Version : sig ... end

val version : unit -> Version.t option

The version at which the current executable was built. The version is None during development, it is only Some _ once the executable is installed or promoted to the source tree.

module Statically_linked_library : sig ... end

module Statically_linked_libraries : sig ... end

All the libraries that where statically linked in